backend/requirements.txt
Flask==2.3.2
Pillow==9.5.0
python-json-logger==2.0.7
Flask-Cors==4.0.1
pytest
pytest-cov
pylint
python-dotenv
boto3